Extract common Tool functions to a separate header
